# apps/backend/src/swagger/openapi.yaml
openapi: 3.0.3
info:
  title: Presentation Builder API
  description: |
    API for managing presentations, projects, slides, and global settings.
    
    ## Base URL
    All endpoints are prefixed with `/api/v1`
    
    ## Response Format
    All successful responses are wrapped in a `data` property:
    ```json
    {
      "data": { ... }
    }
    ```
    
    ## Authentication
    Currently, the API does not require authentication (to be implemented).
    
  version: 1.0.0
  contact:
    name: API Support
  license:
    name: MIT
servers:
  - url: /api/v1
    description: API v1
...
paths:
...
  /upload:
    post:
      tags:
        - Upload
      summary: Upload a media file
      description: |
        Uploads an image or video file to object storage.
        
        **Supported formats:**
        - Images: JPEG, PNG, GIF, WebP, SVG (max 10MB)
        - Videos: MP4, WebM, OGG, QuickTime (max 100MB)
      operationId: uploadMedia
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required:
                - file
              properties:
                file:
                  type: string
                  format: binary
                  description: The media file to upload
      responses:
        '201':
          description: File uploaded successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: '#/components/schemas/UploadResponse'
        '400':
          description: Invalid file or validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
components:
...
  schemas:
...
    UploadResponse:
      type: object
      properties:
        url:
          type: string
          description: Public URL to the uploaded file
          example: 'https://storage.farrel.space/slides/media/abc123.jpg'
        mediaType:
          type: string
          nullable: true
          enum:
            - image
            - video
          description: Detected media type
          example: 'image'
      required:
        - url

    # Error
    Error:
      type: object
      properties:
        statusCode:
          type: integer
          description: HTTP status code
          example: 400
        message:
          type: string
          description: Error message
          example: 'Validation failed'
        error:
          type: string
          description: Error type
          example: 'Bad Request'